Split derytelecomextranetquery.py in 3 files
Separate derytelecomextranetquery.py in 3 files: derytelecomextranetquery.py,
content.py and __main__.py. Place these files in a new subdirectory named
derytelecomextranetquery.

This separation should make the code easier to read and maintain. The new
subdirectory is callable by the python command thanks to the __main__.py file.
This file contain the main code that was in derytelecomextranetquery.py. The
Content classes are now in the content.py file.

Modify the applet.js file so that the call to the python command is made on the
new subdirectory instead of a python file.