Modify DerytelecomExtranetQuery.connect to use PHPSESSID cookie
Modify DerytelecomExtranetQuery.connect to use the value of the PHPSESSID cookie
to generate the name of the input login and the name of the input password
instead of retrieving those names from the html page.

This change should make the program much faster since it has one less html page
to parse.