Add a test after each query to check if we got logged out
Add a test after each query to check if we got logged out of the Derytelecom
Extranet.

Prior to adding this test, DerytelecomExtranetQuery.connect was
not able to detect that a connection had failed because of a bad username or
password. The rest of the script was still executed even if the login process
had failed.

Adding this test to subsequent queries after the login process allow to detect
logouts of the Derytelecom Extranet. These logouts can happen for many
reasons, like when the session expire.

Stopping the python script after a logout is crucial to prevent it
from making subsequent queries to the server.