Add open_in_webbrowser method to DerytelecomExtranetQuery
Add open_in_webbrowser method to the DerytelecomExtranetQuery class. This
method opens the index URL in the web browser and it allows to specify the
query string arguments.

If the module PyUserInput is available and the username and password
parameters are specified, this method will perform an auto login using
keyboard events. I use keyboard events because I wasn't able to find
another way to perform an auto login.