Modify wrap_attributes_in_dict to work with nested classes
Modify the decorator 'wrap_attributes_in_dict' to detect and offer the
possibility to wrap the attributes of nested classes. This change removes the
ability of 'wrap_attributes_in_dict' to get dicts to append as keywords
arguments. This is to avoid confusion with the new keyword argument 'depth'.
This argument allows to control the depth to which nested classes are affected
by this decorator.

This change is to help with the modifications that will be made on the
'PARAM' class of 'derytelecomextranetquery.py'.