Fix 'wrap_attributes_in_dict' so it can work with new style class
Before, to modify the value of an attribute, we modified the dictionary of the
class (stored in the '__dict__' attribute). In new style class, you cannot do
that since the dictionary obtained from '__dict__' is read-only.

To set the new value of the attributes, we now use the function 'setattr'. This
function work with both old and new style classes.