Handle exceptions better in derytelecomextranetquery
An exceptions hierarchy was added to 'derytelecomextranetquery.py'. Now
all exceptions classes of this file are based on the base exception class
'DerytelecomExtranetQueryException'. This hierarchy can be useful for other
modules using this module as it offers them a way to easily catch all the
custom exceptions that can be thrown by this module.

Also, all the custom exceptions of the 'requests' package are now caught and
converted into 'InternetConnectionError'. This allows to hide unnecessary
details to the users of this module. If a 'requests' exception is thrown, it
is most likely due to a problem with the internet connection, that is all the
users need to know.

Some modifications have also been made to the '__main__.py' file. It's now
catching all the exceptions that are thrown by the modules that it uses and it
displays an error message on the console and then return an exit code that is
different for each custom exceptions of 'derytelecomextranetquery.py'. For
other exceptions, it returns a default exit code.