Add new features to the applet
First of all, this commit adds two new spawn functions. Theses functions
replace the one from the Cinnamon 'misc.util' library. The biggest change that
these functions bring is that they pass the exit code of the program that they
executed to the callback. This allows to detect if an error occurred which was
not possible before.

A new feature is now available on this applet. When the user click on the
applet, the Derytelecom Extranet login page opens in the default web browser.
If the new 'Auto-Login' checkbox in this applet configuration is checked, this
applet will try to fill the login form using the provided username and
password. This only works if the PyUserInput package is installed.

Another new feature is that the color of the text displayed by this applet
change depending on whether the last transaction with the Derytelecom Extranet
was successful or not. By default, if it was successful, the text is green else
it is red. These colors can be changed in the configuration of this applet.