Add a comment/docstring at the top of each source files
The comment/docstring describe the content of the file and add some copyright
information.

An encoding declaration was also added at the top of each python source files.
They are all declared with a UTF-8 encoding. This change was necessary because
the special characters in my name were causing some errors.