Make color change in the settings be applied directly
Make color change in the settings be applied directly instead of after the
update delay.

So when a user makes a change to 'Ok color' or 'Error color' in the settings,
the new color will be applied instantly instead of after the update delay. This
change will certainly be more in touch with how the users expect the applet to
react.