Make the installation path more accurate in README.md
The directory 'derytelecomavailabledata@remi.tetreault' was added at the end of
the installation path. If the applet is not in a directory with that name, it
will not work so it was important to clarify. Also, the path is now in a code
block to make it stand out more.